requirements.txt
streamlit
pytesseract
Pillow
requests
beautifulsoup4
